pyproject.toml
[tool.poetry]
name = "homework-24-1"
version = "0.1.0"
description = ""
authors = ["Aleksei Prusikov <imemphatic@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
